Extract section composables to separate files
